+++ src/test/resources/Data.json
@@ -10,7 +10,7 @@
   "policyHeaders": "Returns Policy,Flipkart Terms of Use,Safe and Secure Shopping,Privacy Policy,Sitemap,FLIPKART INDIA PRIVATE LIMITED E-WASTE RECYCLING POLICY",
   "AppKeyHeader": "AppKey",
   "WrongAppKeyValue":"qP2wzibM0y9rLeRc3jAZQEoBMGgtVGj777",
-  "base_url": "https://aws-stg1.api.rccl.com/en/al/web/v3/guestAccounts"
+  "base_url": "https://aws-stg1.api.rccl.com/en/al/web/v3/guestAccounts",
   "AppKeyValue": "qP2wzibM0y9rLeRc3jAZQEoBMGgtVGj7",
   "AppKeyWrongValue": "qP2wzibM0y9rLeRc3jAZQEoBMGgtVGj78",
   "ContentTypeHeader": "Content-Type",
